fix(emit): coerce T22 CPI amount to u64 — bare literals emitted non-compiling code (sweep S7)
transfer_checked(ctx, 1, decimals) emitted `let __t22_amount = (1).to_le_bytes();`
— an untyped integer literal → E0689 'ambiguous numeric type', the .so won't
compile. The emit-time validator marked it clean (blind spot); only cargo
caught it. Fixtures always passed a typed u64 arg, so the 6 T22 checked-amount
sites in pinocchio-emitter were never exercised with a literal.

Wrap each as ((amount) as u64).to_le_bytes() — a no-op cast on typed u64, so
.so bytes (and every byte-equal fixture) are unchanged; re-baselined the
t22-transfer binary-parity snapshot (only the cast text differs).

+++ api/src/emitter/pinocchio-emitter.ts
@@ -953,7 +953,7 @@ export class PinocchioEmitter extends BaseEmitter {
         return `    // Token-2022 transfer (unchecked) — ${from} → ${to}
     {
 ${TOKEN_2022_PROGRAM_ID_CONST}
-        let __t22_amount = (${amount}).to_le_bytes();
+        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 9] = [
             3,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
@@ -1006,7 +1006,7 @@ ${invokeCall}
         : `&TOKEN_2022_PROGRAM_ID`;
       return `    // Token-2022 transfer_checked — ${from} → ${to}
 ${prelude}    {
-${programIdConstBlock}        let __t22_amount = (${amount}).to_le_bytes();
+${programIdConstBlock}        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 10] = [
             12,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
@@ -1051,7 +1051,7 @@ ${invokeCall}
         const invokeCall = emitT22Invoke(`${mint}, ${to}, ${authority}`, signerSeeds);
         return `    // Token-2022 mint_to (unchecked) — ${mint} → ${to}
     {
-${programIdConstBlock}        let __t22_amount = (${amount}).to_le_bytes();
+${programIdConstBlock}        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 9] = [
             7,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
@@ -1076,7 +1076,7 @@ ${invokeCall}
       const invokeCall = emitT22Invoke(`${mint}, ${to}, ${authority}`, signerSeeds);
       return `    // Token-2022 mint_to_checked — ${mint} → ${to}
 ${prelude}    {
-${programIdConstBlock}        let __t22_amount = (${amount}).to_le_bytes();
+${programIdConstBlock}        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 10] = [
             14,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
@@ -1115,7 +1115,7 @@ ${invokeCall}
         const invokeCall = emitT22Invoke(`${from}, ${mint}, ${authority}`, signerSeeds);
         return `    // Token-2022 burn (unchecked) — ${from}
     {
-${programIdConstBlock}        let __t22_amount = (${amount}).to_le_bytes();
+${programIdConstBlock}        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 9] = [
             8,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
@@ -1140,7 +1140,7 @@ ${invokeCall}
       const invokeCall = emitT22Invoke(`${from}, ${mint}, ${authority}`, signerSeeds);
       return `    // Token-2022 burn_checked — ${from}
 ${prelude}    {
-${programIdConstBlock}        let __t22_amount = (${amount}).to_le_bytes();
+${programIdConstBlock}        let __t22_amount = ((${amount}) as u64).to_le_bytes();
         let __t22_data: [u8; 10] = [
             15,
             __t22_amount[0], __t22_amount[1], __t22_amount[2], __t22_amount[3],
